:recycle: Long 타입 ID Integer로 변

+++ src/main/java/com/pricewagon/pricewagon/domain/category/repository/CategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pricewagon.pricewagon.domain.category.entity.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-	List<Category> findAllByParentCategory_Id(Long parentCateghoryId);
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+	List<Category> findAllByParentCategory_Id(Integer parentCateghoryId);
 }
